# Hash tooling changes